+++ src/govActions.sol
@@ -20,8 +20,8 @@ pragma solidity 0.5.15;
 
 interface SetterLike {
     function file(bytes32, address) external;
-    function file(bytes32, uint) external;
-    function file(bytes32, bytes32, uint) external;
+    function file(bytes32, uint256) external;
+    function file(bytes32, bytes32, uint256) external;
     function file(bytes32, bytes32, address) external;
     function rely(address) external;
     function deny(address) external;
@@ -37,7 +37,7 @@ interface EndLike {
 
 interface PauseLike {
     function setAuthority(address) external;
-    function setDelay(uint) external;
+    function setDelay(uint256) external;
 }
 
 contract GovActions {
@@ -45,11 +45,11 @@ contract GovActions {
         SetterLike(who).file(what, data);
     }
 
-    function file(address who, bytes32 what, uint data) public {
+    function file(address who, bytes32 what, uint256 data) public {
         SetterLike(who).file(what, data);
     }
 
-    function file(address who, bytes32 ilk, bytes32 what, uint data) public {
+    function file(address who, bytes32 ilk, bytes32 what, uint256 data) public {
         SetterLike(who).file(ilk, what, data);
     }
 
@@ -57,12 +57,12 @@ contract GovActions {
         SetterLike(who).file(ilk, what, data);
     }
 
-    function dripAndFile(address who, bytes32 what, uint data) public {
+    function dripAndFile(address who, bytes32 what, uint256 data) public {
         SetterLike(who).drip();
         SetterLike(who).file(what, data);
     }
 
-    function dripAndFile(address who, bytes32 ilk, bytes32 what, uint data) public {
+    function dripAndFile(address who, bytes32 ilk, bytes32 what, uint256 data) public {
         SetterLike(who).drip(ilk);
         SetterLike(who).file(ilk, what, data);
     }
@@ -87,11 +87,11 @@ contract GovActions {
         PauseLike(pause).setAuthority(newAuthority);
     }
 
-    function setDelay(address pause, uint newDelay) public {
+    function setDelay(address pause, uint256 newDelay) public {
         PauseLike(pause).setDelay(newDelay);
     }
 
-    function setAuthorityAndDelay(address pause, address newAuthority, uint newDelay) public {
+    function setAuthorityAndDelay(address pause, address newAuthority, uint256 newDelay) public {
         PauseLike(pause).setAuthority(newAuthority);
         PauseLike(pause).setDelay(newDelay);
     }
